requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asyncio==3.4.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.2
curl_cffi==0.6.0b9
dataclasses==0.6
dataclasses-json==0.6.4
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
duckduckgo_search==4.4.3
ecdsa==0.18.0
faiss-cpu==1.7.4
fastapi==0.109.2
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
grpcio==1.60.1
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==2.10
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
injector==0.21.0
install==1.3.5
Jinja2==3.1.3
jose==1.0.0
jsonpatch==1.33
jsonpointer==2.4
jwt==1.3.1
kubernetes==29.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
langsmith==0.0.87
litellm==1.23.16
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pluggy==1.4.0
postgrest==0.15.0
postgrest-py==0.10.6
posthog==3.4.1
protobuf==4.25.2
pulsar-client==3.4.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.2
pyparsing==3.1.1
pypdf==4.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
realtime==1.0.2
realtime-py==0.1.3
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supabase-py==0.0.2
supafunc==0.3.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
